+++ src/main/java/com/banking_rest_api/test_demo_bank/repository/AccountRepository.java
@@ -10,4 +10,4 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Integer> {}
+public interface AccountRepository extends JpaRepository<Account, Long> {}
